requirements.txt
numpy==1.23.3
Pillow==9.4.0
tensorflow==2.11.0
